+++ src/GifPresenter.ts
@@ -1,5 +1,5 @@
 /*!
- gifken v0.5.0
+ gifken
  Copyright (c) 2013 aaharu
  This software is released under the MIT License.
  https://raw.github.com/aaharu/gifken/master/LICENSE